Update BalanceSheetTransformer to make it an interface
- Update BalanceSheetTransformer to make it an interface
- Add BalanceSheetTransformerImpl class to contain concrete implementation
- Update BalanceSheetServiceImpl to use Autowired BalanceSheetTransformer
- Update BalanceSheetServiceImplTests to mock BalanceSheetTransformer
- Update BalanceSheetTransformerImplTests to use interface

Resolves: SFA-43.
