+++ src/main/java/uk/gov/companieshouse/web/accounts/service/smallfull/impl/BalanceSheetServiceImpl.java
@@ -1,5 +1,6 @@
 package uk.gov.companieshouse.web.accounts.service.smallfull.impl;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.companieshouse.web.accounts.model.smallfull.BalanceSheet;
 import uk.gov.companieshouse.web.accounts.service.smallfull.BalanceSheetService;
@@ -8,7 +9,8 @@ import uk.gov.companieshouse.web.accounts.transformer.smallfull.BalanceSheetTran
 @Component
 public class BalanceSheetServiceImpl implements BalanceSheetService {
 
-    private static BalanceSheetTransformer transformer = new BalanceSheetTransformer();
+    @Autowired
+    private BalanceSheetTransformer transformer;
 
     @Override
     public BalanceSheet getBalanceSheet(String companyId, String transactionId, String accountsId) {
