+++ src/main/java/uk/gov/companieshouse/web/accounts/transformer/smallfull/BalanceSheetTransformer.java
@@ -1,15 +1,8 @@
 package uk.gov.companieshouse.web.accounts.transformer.smallfull;
 
 import uk.gov.companieshouse.web.accounts.model.smallfull.BalanceSheet;
-import uk.gov.companieshouse.web.accounts.model.smallfull.CalledUpShareCapitalNotPaid;
 
-public class BalanceSheetTransformer {
+public interface BalanceSheetTransformer {
 
-    public BalanceSheet getBalanceSheet() {
-
-        BalanceSheet balanceSheet = new BalanceSheet();
-        balanceSheet.setCalledUpShareCapitalNotPaid(new CalledUpShareCapitalNotPaid());
-
-        return balanceSheet;
-    }
+    BalanceSheet getBalanceSheet();
 }
