+++ src/main/java/uk/gov/companieshouse/web/accounts/transformer/smallfull/impl/BalanceSheetTransformerImpl.java
@@ -1,0 +1,18 @@
+package uk.gov.companieshouse.web.accounts.transformer.smallfull.impl;
+
+import org.springframework.stereotype.Component;
+import uk.gov.companieshouse.web.accounts.model.smallfull.BalanceSheet;
+import uk.gov.companieshouse.web.accounts.model.smallfull.CalledUpShareCapitalNotPaid;
+import uk.gov.companieshouse.web.accounts.transformer.smallfull.BalanceSheetTransformer;
+
+@Component(value="Foo")
+public class BalanceSheetTransformerImpl implements BalanceSheetTransformer {
+
+    public BalanceSheet getBalanceSheet() {
+
+        BalanceSheet balanceSheet = new BalanceSheet();
+        balanceSheet.setCalledUpShareCapitalNotPaid(new CalledUpShareCapitalNotPaid());
+
+        return balanceSheet;
+    }
+}
