+++ src/test/java/uk/gov/companieshouse/web/accounts/service/smallfull/impl/BalanceSheetServiceImplTests.java
@@ -1,26 +1,41 @@
 package uk.gov.companieshouse.web.accounts.service.smallfull.impl;
 
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.web.accounts.model.smallfull.BalanceSheet;
 import uk.gov.companieshouse.web.accounts.service.smallfull.BalanceSheetService;
+import uk.gov.companieshouse.web.accounts.transformer.smallfull.BalanceSheetTransformer;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.mockito.Mockito.when;
 
+@ExtendWith(MockitoExtension.class)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class BalanceSheetServiceImplTests {
 
-    BalanceSheetService balanceSheetService;
+    @Mock
+    private BalanceSheetTransformer transformer;
+
+    @InjectMocks
+    private BalanceSheetService balanceSheetService = new BalanceSheetServiceImpl();
 
     @BeforeEach
     private void init() {
-        balanceSheetService = new BalanceSheetServiceImpl();
+
+        when(transformer.getBalanceSheet()).thenReturn(new BalanceSheet());
     }
 
     @Test
     public void getBalanceSheetSuccess() {
 
-        BalanceSheet balanceSheet = balanceSheetService.getBalanceSheet(any(String.class), any(String.class), any(String.class));
+        BalanceSheet balanceSheet = balanceSheetService.getBalanceSheet("", "", "");
+
         assertNotNull(balanceSheet);
     }
 }
