+++ src/test/java/uk/gov/companieshouse/web/accounts/transformer/smallfull/BalanceSheetTransformerTests.java
@@ -1,20 +1,14 @@
 package uk.gov.companieshouse.web.accounts.transformer.smallfull;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import uk.gov.companieshouse.web.accounts.model.smallfull.BalanceSheet;
+import uk.gov.companieshouse.web.accounts.transformer.smallfull.impl.BalanceSheetTransformerImpl;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class BalanceSheetTransformerTests {
 
-    private BalanceSheetTransformer transformer;
-
-    @BeforeEach
-    public void init() {
-
-        transformer = new BalanceSheetTransformer();
-    }
+    private BalanceSheetTransformer transformer = new BalanceSheetTransformerImpl();
 
     @Test
     public void getBalanceSheetSuccess() {
